Move the PlayQueue implementaion to shared
The file js/app/services/playqueueservice.ts now contains only the Angular
registration of the service while all the logic of the service can be found
from the class definition in js/shared/playqueue.ts.
